infra/requirements.txt
# AWS CDK v2 - Latest stable with Python 3.11
aws-cdk-lib==2.150.0
constructs==10.3.0

# Additional dependencies for testing and development
pytest
pytest-cov
boto3
moto